+++ src/main/java/org/folio/roles/repository/CapabilityRepository.java
@@ -346,6 +346,11 @@ public interface CapabilityRepository extends BaseCqlJpaRepository<CapabilityEnt
     order by entity.name""")
   List<CapabilityEntity> findAllByPermissionNames(@Param("names") Collection<String> names);
 
+  @Query("""
+    select entity from CapabilityEntity entity where entity.permission in :names
+    order by entity.name""")
+  List<CapabilityEntity> findAllByPermissionNamesIncludeDummy(@Param("names") Collection<String> names);
+
   @Query(nativeQuery = true, value = """
     WITH prefixes AS (
       SELECT prefix || '%' AS pattern