+++ NEWS.md
@@ -3,6 +3,7 @@
 * Introduce configuration for FSSP (APPPOCTOOL-59)
 * Provide "Data Import" role with permission for central tenant record update through data import (MODROLESKC-304)
 * Fix missing permission for loadable role, fix duplicated capabilities by permission name (MODROLESKC-313)
+* Implement replacement for dummy capabilities (MODROLESKC-306)
 
 ## Version `v3.0.0` (14.03.2025)
 * Error with roles migration (MODROLESKC-282)